♻️ extract error.types
