+++ packages/core/src/domain/error/trackRuntimeError.ts
@@ -1,8 +1,9 @@
 import type { Observable } from '../../tools/observable'
 import { clocksNow } from '../../tools/utils/timeUtils'
 import { startUnhandledErrorCollection } from '../tracekit'
-import { ErrorSource, computeRawError, ErrorHandling } from './error'
-import type { RawError } from './error'
+import { computeRawError } from './error'
+import type { RawError } from './error.types'
+import { ErrorHandling, ErrorSource } from './error.types'
 
 export function trackRuntimeError(errorObservable: Observable<RawError>) {
   return startUnhandledErrorCollection((stackTrace, originalError) => {
