+++ packages/core/src/domain/eventRateLimiter/createEventRateLimiter.spec.ts
@@ -1,9 +1,9 @@
 import type { Clock } from '../../../test'
 import { mockClock } from '../../../test'
-import type { RawError } from '../error/error'
 import type { RelativeTime } from '../../tools/utils/timeUtils'
 import { relativeToClocks, resetNavigationStart, ONE_MINUTE } from '../../tools/utils/timeUtils'
 import { noop } from '../../tools/utils/functionUtils'
+import type { RawError } from '../error/error.types'
 import { createEventRateLimiter } from './createEventRateLimiter'
 import type { EventRateLimiter } from './createEventRateLimiter'
