+++ packages/core/src/transport/httpRequest.ts
@@ -2,8 +2,8 @@ import type { EndpointBuilder } from '../domain/configuration'
 import { addTelemetryError } from '../domain/telemetry'
 import type { Context } from '../tools/serialisation/context'
 import { monitor } from '../tools/monitor'
-import type { RawError } from '../domain/error/error'
 import { addEventListener } from '../browser/addEventListener'
+import type { RawError } from '../domain/error/error.types'
 import { newRetryState, sendWithRetryStrategy } from './sendWithRetryStrategy'
 import type { FlushReason } from './flushController'
