+++ packages/core/src/transport/startBatchWithReplica.ts
@@ -1,8 +1,8 @@
 import type { Configuration, EndpointBuilder } from '../domain/configuration'
-import type { RawError } from '../domain/error/error'
 import type { Context } from '../tools/serialisation/context'
 import type { Observable } from '../tools/observable'
 import type { PageExitEvent } from '../browser/pageExitObservable'
+import type { RawError } from '../domain/error/error.types'
 import { Batch } from './batch'
 import { createHttpRequest } from './httpRequest'
 import { createFlushController } from './flushController'
